requirements.txt
aiofiles==23.1.0
alembic==1.10.3
discord.py==2.2.2
pydantic==1.10.7
python-dotenv==1.0.0
sqlalchemy==2.0.10
tzdata
uvloop==0.17.0; sys_platform == 'linux'
